+++ src/cli/ArgumentsBuilder.php
@@ -31,6 +31,7 @@ final readonly class ArgumentsBuilder
                 $argv,
                 'hv',
                 [
+                    'paths=',
                     'help',
                     'version',
                 ],
@@ -45,6 +46,7 @@ final readonly class ArgumentsBuilder
             // @codeCoverageIgnoreEnd
         }
 
+        $paths   = null;
         $help    = false;
         $version = false;
 
@@ -52,6 +54,13 @@ final readonly class ArgumentsBuilder
             assert(is_array($option));
 
             switch ($option[0]) {
+                case '--paths':
+                    $paths = (string) $option[1];
+
+                    assert($paths !== '');
+
+                    break;
+
                 case 'h':
                 case '--help':
                     $help = true;
@@ -68,6 +77,7 @@ final readonly class ArgumentsBuilder
 
         return new Arguments(
             $options[1],
+            $paths,
             $help,
             $version,
         );